requirements.txt
flask
gunicorn
python-binance
logging

